Exercise2/backend/requirements.txt
chromadb
ollama
fastapi
mysql-connector-python